msm: camera: Fix camera recording corruption on isp32
For ISP write master configuration, the buffer height
should be configured as image height instead of buffer
scanlines because for cases where scanlines is larger than
image height, bottom padding can be ignored by
not writing to the buffer. In current configuration,
image data of the next frame will be written to padding,
which result in image corruption.

Change-Id: Ibb8416e27d9f77a3181dc8793743d6b0c1661126
Signed-off-by: Kevin Chan <ktchan@codeaurora.org>
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>