msm: camera: Handle burst stream with limited buffer
For burst stream, image buffers are not queued back
for processing. When ISP ping pong address is updated,
we need to check for the remaining burst count to see
if we need to fetch and new buffer and update the address.
Add change to update the burst count in order to check if
we have free buffers.

Change-Id: I683889d6e04de59ef610a5abb771b103fd4669f2
Signed-off-by: Kevin Chan <ktchan@codeaurora.org>